Move discovery components in utils
