+++ lib/ui/utils/discovery_components.dart
